+++ raytracer/src/render/scene.rs
@@ -16,7 +16,7 @@ pub struct Scene<'a, const NUM_COMPONENTS: usize> {
 unsafe impl<const NUM_COMPONENTS: usize> Send for Scene<'_, NUM_COMPONENTS> {}
 
 impl<const NUM_COMPONENTS: usize> Hittable for Scene<'_, NUM_COMPONENTS> {
-    fn hit(&self, r: &Ray, dist_range: (f64, f64)) -> Option<HitData> {
+    fn hit(&self, r: &Ray, dist_range: (f32, f32)) -> Option<HitData> {
         let mut closest_so_far = dist_range.1;
         let mut hit_data_option: Option<HitData> = None;
