+++ raytracer/src/render/shapes/hit.rs
@@ -6,11 +6,11 @@ use crate::{
 pub struct HitData<'a> {
     pub p: Point3,
     pub normal: Vec3,
-    pub dist: f64,
+    pub dist: f32,
     pub front_face: bool,
     pub material: &'a Material,
 }
 
 pub trait Hittable {
-    fn hit(&self, r: &Ray, dist_range: (f64, f64)) -> Option<HitData>;
+    fn hit(&self, r: &Ray, dist_range: (f32, f32)) -> Option<HitData>;
 }
